Restructure modules and add closed-form formula
