+++ src/ambulance_game/markov/__init__.py
@@ -16,15 +16,25 @@ from .markov import (
     get_mean_waiting_time_markov,
 )
 
-from .additional import (
-    generate_code_for_tikz_figure,
-    build_body_of_tikz_spanning_tree,
+from .graphical import (
     reset_L_and_R_in_array,
     find_next_permutation_over,
     find_next_permutation_over_L_and_R,
     generate_next_permutation_of_edges,
     check_permutation_is_valid,
+    get_rate_of_state_00_graphically,
+    get_all_permutations,
+    get_permutations_ending_in_R,
+    get_permutations_ending_in_D_where_any_RL_exists,
+    get_permutations_ending_in_L_where_any_RL_exists,
+    get_permutations_ending_in_RL_where_RL_exists_only_at_the_end,
+    get_coefficient,
+)
+
+
+from .tikz import(
     get_tikz_code_for_permutation,
     generate_code_for_tikz_spanning_trees_rooted_at_00,
-    get_rate_of_state_00_graphically,
+    generate_code_for_tikz_figure,
+    build_body_of_tikz_spanning_tree,
 )
